Migrate checking response status code to CodeVerified component
